requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
requests
python-multipart